Add certs + some fixes

+++ src/main/java/com/sili/model/AValueTO.java
@@ -12,5 +12,5 @@ public class AValueTO {
 
     Long x;
     List<Integer> a;
-    Long userId;
+    Integer userId;
 }
